requirements.txt
Django==5.2.1
Pillow==11.2.1
torch>=2.0.0
diffusers>=0.24.0
transformers>=4.35.0
peft>=0.6.0
accelerate>=0.24.0
safetensors>=0.4.0
numpy>=1.24.0
opencv-python>=4.8.0
requests>=2.31.0 